fix(06-uat): fix 3 bugs found during UAT + complete phase 6 verification
- strip /works/ prefix from olKey in all DTO responses so frontend URLs
  are clean path segments (was causing Tomcat 400 on book detail page)
- add @Transactional(readOnly=true) to ShelfService.getEntry() to fix
  LazyInitializationException on GET /shelf/:id (500 for all entries)
- fix ShelfPage card navigation: /shelf/:id → /shelf/:id/edit
- fix pre-existing ShelfServiceTest rating Short/Integer type mismatch
- update BookControllerTest, BookServiceTest, BookCacheIntegrationTest
  assertions to expect bare olKey form (OL45804W not /works/OL45804W)
- add 06-UAT.md: 15/16 passed, 1 skipped (PWA install prompt / HTTPS)

+++ src/test/java/com/booktracker/shelf/ShelfServiceTest.java
@@ -176,7 +176,7 @@ class ShelfServiceTest {
 
         shelfService.updateMetadata(entryId, req, user);
 
-        assertThat(entry.getRating()).isEqualTo(4);
+        assertThat(entry.getRating()).isEqualTo((short) 4);
     }
 
     // ----------------------------------------------------------------
